requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
pip-audit==2.7.2
Flask-SQLAlchemy==3.1.1
email-validator
PyMySQL==1.1.0
